Moved out store content into designated files
