+++ src/store/modules/splits.module.ts
@@ -1,10 +1,10 @@
 import { Injector } from 'lightweight-di';
 
-import { Module } from '..';
 import { SplitsFile } from '../../models/files';
 import { VALIDATOR_SERVICE_TOKEN } from '../../models/services';
 import { Segment, TimingMethod } from '../../models/splits';
 import { SplitsState } from '../../models/states/splits.state';
+import { Module } from '../../models/store';
 import { asCleanNumber } from '../../utils/converters';
 import { getFinalTime } from '../../utils/time';
