+++ src/utils/keys.ts
@@ -31,10 +31,10 @@ export function keyToAcceleratorString(keyCode: string) {
                     return 'Super';
             }
         default:
-        if (keycode.names[keyCode] != null) {
-            return keycode.names[keyCode].toUpperCase();
-        } else {
-            return keyCode.toUpperCase();
-        }
+            if (keycode.names[keyCode] != null) {
+                return keycode.names[keyCode].toUpperCase();
+            } else {
+                return keyCode.toUpperCase();
+            }
     }
 }
