+++ src/vue.d.ts
@@ -1,9 +1,9 @@
 import { Injector } from 'lightweight-di';
+import { WatchOptions } from 'vue';
 import Vue from 'vuex';
 
 import { RootState } from './models/states/root.state';
-import { Commit } from './store';
-import { WatchOptions } from 'vue';
+import { Commit } from './models/store';
 
 type StoreObserveHandler = (state: RootState) => any;
