+++ tests/mocks/store-mock.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from 'lightweight-di';
 
 import { StoreInterface } from '../../src/models/services';
 import { RootState } from '../../src/models/states/root.state';
-import { Commit } from '../../src/store';
+import { Commit } from '../../src/models/store';
 import { createCommit } from '../../src/utils/store';
 
 @Injectable
